# openapi.yaml
openapi: 3.0.0
paths:
...
  /orders:
    post:
      operationId: OrderController_create
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateOrderDto"
      responses:
        "201":
          description: ""
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OrderResponse"
      security: &a14
        - bearer: []
      summary: Create an order
      tags: &a15
        - Orders
    get:
      operationId: OrderController_list
      parameters:
        - name: page
          required: false
          in: query
          schema:
            minimum: 1
            default: 1
            type: number
        - name: limit
          required: false
          in: query
          schema:
            minimum: 1
            maximum: 100
            default: 20
            type: number
        - name: q
          required: false
          in: query
          description: Free-text search query.
          schema:
            type: string
        - name: status
          required: false
          in: query
          schema:
            type: string
            enum:
              - pending
              - completed
              - cancelled
        - name: customerId
          required: false
          in: query
          schema:
            type: string
      responses:
        "200":
          description: ""
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PaginatedOrderResponse"
      security: *a14
      summary: List orders (paginated, searchable)
      tags: *a15
  /orders/{id}:
    get:
      operationId: OrderController_getById
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      responses:
        "200":
          description: ""
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OrderResponse"
      security: *a14
      summary: Get an order by id
      tags: *a15
    patch:
      operationId: OrderController_updateItems
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateOrderDto"
      responses:
        "200":
          description: ""
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OrderResponse"
      security: *a14
      summary: Edit the items of a pending order
      tags: *a15
  /orders/{id}/status:
    patch:
      operationId: OrderController_updateStatus
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateOrderStatusDto"
      responses:
        "200":
          description: ""
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OrderResponse"
      security: *a14
      summary: Update an order's status
      tags: *a15
...
info:
  title: Ventura API
  description: Ventura backend API.
  version: 1.0.0
  contact: {}
...
servers: []
components:
  securitySchemes:
    bearer:
      scheme: bearer
      bearerFormat: JWT
      type: http
      description: JWT access token from /auth sign-in.
  schemas:
...
    PaginationMeta:
      type: object
      properties:
        total:
          type: number
          example: 137
        page:
          type: number
          example: 1
        limit:
          type: number
          example: 20
        totalPages:
          type: number
          example: 7
      required:
        - total
        - page
        - limit
        - totalPages
...
    CreateOrderItemDto:
      type: object
      properties:
        resourceId:
          type: string
          example: 665f1b2c3d4e5f6a7b8c9d0e
        quantity:
          type: number
          example: 2
      required:
        - resourceId
        - quantity
    CreateOrderDto:
      type: object
      properties:
        customerId:
          type: string
          example: 665f1b2c3d4e5f6a7b8c9d0e
        items:
          type: array
          items:
            $ref: "#/components/schemas/CreateOrderItemDto"
      required:
        - customerId
        - items
    OrderItemResponse:
      type: object
      properties:
        resourceId:
          type: string
          example: 665f1b2c3d4e5f6a7b8c9d0e
        type:
          type: string
          enum:
            - product
            - service
          example: product
        name:
          type: string
          example: Espresso Beans 1kg
        price:
          type: number
          example: 9.99
        quantity:
          type: number
          example: 2
        subTotal:
          type: number
          example: 19.98
      required:
        - resourceId
        - type
        - name
        - price
        - quantity
        - subTotal
    OrderResponse:
      type: object
      properties:
        id:
          type: string
          example: 665f1b2c3d4e5f6a7b8c9d0e
        orderNumber:
          type: string
          example: ORD-AB3XY9KP
        businessId:
          type: string
          example: 665f1b2c3d4e5f6a7b8c9d0e
        customerId:
          type: string
          example: 665f1b2c3d4e5f6a7b8c9d0e
        customerName:
          type: string
          example: Ada Lovelace
        customerEmail:
          type: object
          nullable: true
          example: ada@example.com
          format: email
        customerPhone:
          type: object
          nullable: true
          example: "+15551234567"
        items:
          type: array
          items:
            $ref: "#/components/schemas/OrderItemResponse"
        totalAmount:
          type: number
          example: 19.98
        status:
          type: string
          enum:
            - pending
            - completed
            - cancelled
          example: pending
        invoiceId:
          type: object
          nullable: true
          example: 665f1b2c3d4e5f6a7b8c9d0e
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required:
        - id
        - orderNumber
        - businessId
        - customerId
        - customerName
        - items
        - totalAmount
        - status
        - createdAt
        - updatedAt
    PaginatedOrderResponse:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/OrderResponse"
        meta:
          $ref: "#/components/schemas/PaginationMeta"
      required:
        - data
        - meta
    UpdateOrderDto:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/CreateOrderItemDto"
      required:
        - items
    UpdateOrderStatusDto:
      type: object
      properties:
        status:
          type: string
          enum:
            - pending
            - completed
            - cancelled
          example: completed
      required:
        - status
...